requirements.txt
fastapi
PyPDF2
python-dotenv
transformers
uvicorn
torch
numpy
einops
streamlit
pydantic_settings
pdfminer.six